move 2 component dirs move details and student list component
