+++ src/pages/AkademikMahasiswa.Mahasiswa.Detail/index.js
