+++ src/pages/AkademikMahasiswa/index.js
@@ -4,20 +4,11 @@ import { RootProvider } from 'pages/Root/hoc'
 import { useRouteMatch } from 'react-router'
 import Students from './Students'
 import Curriculums from './Curriculums'
-import Generations from './Generations'
 
 export const AkademikMahasiswa = () => {
   const { path } = useRouteMatch();
   document.title = "Dashboard - Akademik Kemahasiswaan"
   const navigation = useMemo(() => ([
-    {
-      "title": "Halaman Utama",
-      "text": "Halaman Utama",
-      "component": Generations,
-      "path": `${path}`,
-      exact: true,
-      icon: 'home'
-    },
     {
       "title": "Mahasiswa",
       "text": "Mahasiswa",
